Extracted shared logic to BaseRequestDetailsActivity
